FCFS Summary Chart Generation

+++ Assets/src/SummaryData.cs
@@ -4,7 +4,7 @@ public class SummaryData
     public string ProcessName;
     public int ArrivalTime;
     public int BurstTime;
-    public int CompilationTime;
-    public int TurnAroundTime;
-    public int WaitingTime;
+    public float CompilationTime;
+    public float TurnAroundTime;
+    public float WaitingTime;
 }
